Refactor type-specific stats handling into runtime types
Refactors StatsFormatter to remove code that's specific to captures,
derivations, and materializations. This code is moved into the specific runtime
files. Also removes loading of task specs from task_term.go and back into the
specific runtime files.
