+++ go/runtime/task_term.go
@@ -61,7 +61,6 @@ func (t *taskTerm) initTerm(shard consumer.Shard, host *FlowConsumer) error {
 		return err
 	}
 
-	var taskSpec pf.Task
 	var statsCollectionSpec *pf.CollectionSpec
 	var logsCollectionSpec *pf.CollectionSpec
 	if err = t.build.Extract(func(db *sql.DB) error {
@@ -70,10 +69,6 @@ func (t *taskTerm) initTerm(shard consumer.Shard, host *FlowConsumer) error {
 		}
 		if statsCollectionSpec, err = catalog.LoadCollection(db, statsCollection(t.labels.TaskName).String()); err != nil {
 			return fmt.Errorf("loading stats collection: %w", err)
-		}
-
-		if taskSpec, err = loadTask(db, t.labels.TaskType, t.labels.TaskName); err != nil {
-			return fmt.Errorf("loading task spec: %w", err)
 		}
 
 		return nil
@@ -95,19 +90,14 @@ func (t *taskTerm) initTerm(shard consumer.Shard, host *FlowConsumer) error {
 	if t.StatsFormatter, err = NewStatsFormatter(
 		t.labels,
 		statsCollectionSpec,
-		taskSpec,
 	); err != nil {
 		return err
 	}
 
-	var logFields = log.Fields{
+	t.Log(log.InfoLevel, log.Fields{
 		"labels":     t.labels,
 		"lastLabels": lastLabels,
-	}
-	if t.LogPublisher.Level() >= log.DebugLevel {
-		logFields["taskSpec"] = taskSpec
-	}
-	t.Log(log.InfoLevel, logFields, "initialized catalog task term")
+	}, "initialized catalog task term")
 
 	return nil
 }
@@ -210,18 +200,4 @@ func signalOnSpecUpdate(ks *keyspace.KeySpace, shard consumer.Shard, spec *pf.Sh
 			return
 		}
 	}
-}
-
-func loadTask(db *sql.DB, taskType string, taskName string) (task pf.Task, err error) {
-	switch taskType {
-	case labels.TaskTypeCapture:
-		task, err = catalog.LoadCapture(db, taskName)
-	case labels.TaskTypeDerivation:
-		task, err = catalog.LoadDerivation(db, taskName)
-	case labels.TaskTypeMaterialization:
-		task, err = catalog.LoadMaterialization(db, taskName)
-	default:
-		err = fmt.Errorf("invalid task type '%s' for task: '%s'", taskType, taskName)
-	}
-	return
 }
